+++ src/sll/error.c
@@ -40,12 +40,6 @@ __SLL_FUNC void sll_print_error(sll_input_data_stream_t* is,const sll_error_t* e
 			case SLL_ERROR_INVALID_STACK_INDEX:
 				printf("Stack Index Out of Bounds\n");
 				return;
-			case SLL_ERROR_UNRELEASED_OBJECTS:
-				printf("Unreleased Runtime Objects\n");
-				return;
-			case SLL_ERROR_ASSERTION:
-				printf("%s\n",e->dt.str);
-				return;
 		}
 	}
 	sll_file_offset_t os=e->dt.r.off;