+++ tests/run_tests.c
@@ -108,7 +108,6 @@ char* error_to_string(sll_error_t* e){
 		case SLL_ERROR_STACK_CORRUPTED:
 		case SLL_ERROR_INVALID_INSTRUCTION_INDEX:
 		case SLL_ERROR_INVALID_STACK_INDEX:
-		case SLL_ERROR_UNRELEASED_OBJECTS:
 			snprintf(o,512,"<type=%"PRIu8">",e->t);
 			break;
 		case SLL_ERROR_INTERNAL_STACK_OVERFLOW:
@@ -145,7 +144,6 @@ char* error_to_string(sll_error_t* e){
 			snprintf(o,512,"<type=%"PRIu8", range=%"PRIu32"-%"PRIu32">",e->t,e->dt.r.off,e->dt.r.off+e->dt.r.sz);
 			break;
 		case SLL_ERROR_UNKNOWN_INTERNAL_FUNCTION:
-		case SLL_ERROR_ASSERTION:
 			snprintf(o,512,"<type=%"PRIu8", string='%s'>",e->t,e->dt.str);
 			break;
 		case SLL_ERROR_INVALID_INSTRUCTION:
@@ -339,7 +337,6 @@ void run_parser_test(const char* fp,test_result_t* o){
 				ne.dt.r.sz=(sll_file_offset_t)(err_v_e->dt.a.dt+1)->dt.i-ne.dt.r.off;
 				break;
 			case SLL_ERROR_UNKNOWN_INTERNAL_FUNCTION:
-			case SLL_ERROR_ASSERTION:
 				if (!err_v_e||err_v_e->t!=SLL_JSON_OBJECT_TYPE_STRING||err_v_e->dt.s.l>255){
 					o->s++;
 					printf("-> JSON Error in Test Case #%"PRIu32"\n",i);
@@ -406,7 +403,6 @@ _wrong_error:
 				}
 				break;
 			case SLL_ERROR_UNKNOWN_INTERNAL_FUNCTION:
-			case SLL_ERROR_ASSERTION:
 				if (strcmp((char*)ne.dt.str,(char*)e->dt.str)){
 					goto _wrong_error;
 				}